moved ship module dialog out of ship info
